+++ jason/src/peer/mod.rs
@@ -511,7 +511,8 @@ impl PeerConnection {
     /// Tells the [`RtcPeerConnection`] to start ICE restart on renegotiation.
     ///
     /// After this function will be called, the offer returned by the next call
-    /// to createOffer() is automatically configured to trigger ICE restart.
+    /// to [`PeerConnection::get_offer`] is automatically configured to trigger
+    /// ICE restart.
     pub fn restart_ice(&self) {
         self.peer.restart_ice();
     }