+++ src/media/peer.rs
@@ -33,8 +33,8 @@
 //! # Implementing [`Peer`]'s update that requires (re)negotiation
 //!
 //! 1. All changes that require (re)negotiation should be done by adding a new
-//!    variant into [`TrackChange`].
-//! 2. Implement your changing logic in the [`TrackChangeHandler`]
+//!    variant into [`PeerChange`].
+//! 2. Implement your changing logic in the [`PeerChangeHandler`]
 //!    implementation.
 //! 3. Create a function in the [`PeerChangesScheduler`] which will schedule
 //!    your change by adding it into the [`Context::track_changes_queue`].
@@ -81,7 +81,7 @@ pub trait PeerUpdatesSubscriber: fmt::Debug {
     /// Notifies subscriber that provided [`Peer`] must be negotiated.
     fn negotiation_needed(&self, peer_id: PeerId);
 
-    /// Notifies subscriber that provided [`TrackUpdate`] were forcibly (without
+    /// Notifies subscriber that provided [`PeerUpdate`] were forcibly (without
     /// negotiation) applied to [`Peer`].
     fn force_update(&self, peer_id: PeerId, changes: Vec<PeerUpdate>);
 }
@@ -328,12 +328,12 @@ pub struct Context {
     /// Indicator whether this [`Peer`] was created on remote.
     is_known_to_remote: bool,
 
-    /// Tracks changes, that remote [`Peer`] is not aware of.
-    pending_track_updates: Vec<TrackChange>,
+    /// Changes, that remote [`Peer`] is not aware of.
+    pending_peer_updates: Vec<PeerChange>,
 
-    /// Queue of the [`TrackChange`]s that are scheduled to apply when this
+    /// Queue of the [`PeerChange`]s that are scheduled to apply when this
     /// [`Peer`] will be in a [`Stable`] state.
-    track_changes_queue: VecDeque<TrackChange>,
+    peer_changes_queue: VecDeque<PeerChange>,
 
     /// Subscriber to the events which indicates that negotiation process
     /// should be started for this [`Peer`].
@@ -343,7 +343,7 @@ pub struct Context {
 /// Tracks changes, that remote [`Peer`] is not aware of.
 #[dispatchable]
 #[derive(Clone, Debug)]
-enum TrackChange {
+enum PeerChange {
     /// [`MediaTrack`]s with [`Direction::Send`] of this [`Peer`] that remote
     /// Peer is not aware of.
     AddSendTrack(Rc<MediaTrack>),
@@ -359,10 +359,10 @@ enum TrackChange {
     IceRestart,
 }
 
-impl TrackChange {
-    /// Tries to return new [`Track`] based on this [`TrackChange`].
+impl PeerChange {
+    /// Tries to return new [`Track`] based on this [`PeerChange`].
     ///
-    /// Returns `None` if this [`TrackChange`] doesn't indicates new [`Track`]
+    /// Returns `None` if this [`PeerChange`] doesn't indicates new [`Track`]
     /// creation.
     fn as_new_track(&self, partner_member_id: MemberId) -> Option<Track> {
         match self.as_track_update(partner_member_id) {
@@ -371,10 +371,10 @@ impl TrackChange {
         }
     }
 
-    /// Returns [`TrackUpdate`] based on this [`TrackChange`].
+    /// Returns [`PeerUpdate`] based on this [`PeerChange`].
     fn as_track_update(&self, partner_member_id: MemberId) -> PeerUpdate {
         match self {
-            TrackChange::AddSendTrack(track) => PeerUpdate::Added(Track {
+            PeerChange::AddSendTrack(track) => PeerUpdate::Added(Track {
                 id: track.id,
                 media_type: track.media_type.clone(),
                 direction: Direction::Send {
@@ -382,7 +382,7 @@ impl TrackChange {
                     mid: track.mid(),
                 },
             }),
-            TrackChange::AddRecvTrack(track) => PeerUpdate::Added(Track {
+            PeerChange::AddRecvTrack(track) => PeerUpdate::Added(Track {
                 id: track.id,
                 media_type: track.media_type.clone(),
                 direction: Direction::Recv {
@@ -390,25 +390,25 @@ impl TrackChange {
                     mid: track.mid(),
                 },
             }),
-            TrackChange::TrackPatch(track_patch) => {
+            PeerChange::TrackPatch(track_patch) => {
                 PeerUpdate::Updated(track_patch.clone())
             }
-            TrackChange::IceRestart => PeerUpdate::IceRestart,
+            PeerChange::IceRestart => PeerUpdate::IceRestart,
         }
     }
 
-    /// Returns `true` if this [`TrackChange`] can be forcibly applied.
+    /// Returns `true` if this [`PeerChange`] can be forcibly applied.
     fn can_force_apply(&self) -> bool {
         match self {
-            TrackChange::AddSendTrack(_)
-            | TrackChange::AddRecvTrack(_)
-            | TrackChange::IceRestart => false,
-            TrackChange::TrackPatch(_) => true,
+            PeerChange::AddSendTrack(_)
+            | PeerChange::AddRecvTrack(_)
+            | PeerChange::IceRestart => false,
+            PeerChange::TrackPatch(_) => true,
         }
     }
 }
 
-impl<T> TrackChangeHandler for Peer<T> {
+impl<T> PeerChangeHandler for Peer<T> {
     type Output = ();
 
     /// Inserts provided [`MediaTrack`] into [`Context::senders`].
@@ -470,11 +470,11 @@ impl<T> Peer<T> {
         self.context.partner_member.clone()
     }
 
-    /// Returns [`TrackUpdate`]s of this [`Peer`] which should be sent to the
-    /// client in the [`Event::TracksApplied`].
+    /// Returns [`PeerUpdate`]s of this [`Peer`] which should be sent to the
+    /// client in the [`Event::PeerUpdated`].
     pub fn get_updates(&self) -> Vec<PeerUpdate> {
         self.context
-            .pending_track_updates
+            .pending_peer_updates
             .iter()
             .map(|c| c.as_track_update(self.partner_member_id()))
             .collect()
@@ -483,7 +483,7 @@ impl<T> Peer<T> {
     /// Returns [`Track`]s that remote [`Peer`] is not aware of.
     pub fn new_tracks(&self) -> Vec<Track> {
         self.context
-            .pending_track_updates
+            .pending_peer_updates
             .iter()
             .filter_map(|c| c.as_new_track(self.partner_member_id()))
             .collect()
@@ -544,11 +544,11 @@ impl<T> Peer<T> {
         &self.context.senders
     }
 
-    /// Commits all [`TrackChange`]s which are marked as forcible
-    /// ([`TrackChange::can_force_apply`]).
+    /// Commits all [`PeerChange`]s which are marked as forcible
+    /// ([`PeerChange::can_force_apply`]).
     pub fn inner_force_commit_scheduled_changes(&mut self) {
         let track_changes_queue = std::mem::replace(
-            &mut self.context.track_changes_queue,
+            &mut self.context.peer_changes_queue,
             VecDeque::new(),
         );
         let mut forcible_changes = VecDeque::new();
@@ -560,7 +560,7 @@ impl<T> Peer<T> {
                 filtered_changes_queue.push_back(track_change);
             }
         }
-        self.context.track_changes_queue = filtered_changes_queue;
+        self.context.peer_changes_queue = filtered_changes_queue;
 
         let mut updates = Vec::new();
         for change in forcible_changes {
@@ -569,8 +569,7 @@ impl<T> Peer<T> {
             updates.push(track_update);
         }
 
-        self.dedup_ice_restarts();
-        self.dedup_track_patches();
+        self.dedup_peer_changes();
 
         if !updates.is_empty() {
             self.context
@@ -595,19 +594,25 @@ impl<T> Peer<T> {
         }
     }
 
-    /// Dedups [`TrackChange::IceRestart`]s.
+    /// Dedups [`PeerChange`]s from this [`Peer`].
+    fn dedup_peer_changes(&mut self) {
+        self.dedup_ice_restarts();
+        self.dedup_track_patches();
+    }
+
+    /// Dedups [`PeerChange::IceRestart`]s.
     fn dedup_ice_restarts(&mut self) {
         let ice_restarts_count = self
             .context
-            .pending_track_updates
+            .pending_peer_updates
             .iter()
-            .filter(|c| matches!(c, TrackChange::IceRestart))
+            .filter(|c| matches!(c, PeerChange::IceRestart))
             .count();
         if ice_restarts_count > 1 {
             let ice_restart_to_remove_count = ice_restarts_count - 1;
             let mut removed_ice_restarts = 0;
-            self.context.pending_track_updates.retain(|c| {
-                if let TrackChange::IceRestart = c {
+            self.context.pending_peer_updates.retain(|c| {
+                if let PeerChange::IceRestart = c {
                     if removed_ice_restarts < ice_restart_to_remove_count {
                         removed_ice_restarts += 1;
                         true
@@ -621,15 +626,15 @@ impl<T> Peer<T> {
         }
     }
 
-    /// Deduplicates pending [`TrackChanges`]s.
+    /// Deduplicates pending [`PeerChanges`]s.
     fn dedup_track_patches(&mut self) {
         let mut grouped_patches: HashMap<TrackId, TrackPatch> = HashMap::new();
         let mut track_changes = Vec::new();
 
         for track_change in
-            std::mem::take(&mut self.context.pending_track_updates)
+            std::mem::take(&mut self.context.pending_peer_updates)
         {
-            if let TrackChange::TrackPatch(patch) = track_change {
+            if let PeerChange::TrackPatch(patch) = track_change {
                 grouped_patches
                     .entry(patch.id)
                     .or_insert_with(|| TrackPatch::new(patch.id))
@@ -642,10 +647,10 @@ impl<T> Peer<T> {
         track_changes.extend(
             grouped_patches
                 .into_iter()
-                .map(|(_, patch)| TrackChange::TrackPatch(patch)),
+                .map(|(_, patch)| PeerChange::TrackPatch(patch)),
         );
 
-        self.context.pending_track_updates = track_changes;
+        self.context.pending_peer_updates = track_changes;
     }
 }
 
@@ -772,8 +777,8 @@ impl Peer<Stable> {
             is_force_relayed,
             endpoints: Vec::new(),
             is_known_to_remote: false,
-            pending_track_updates: Vec::new(),
-            track_changes_queue: VecDeque::new(),
+            pending_peer_updates: Vec::new(),
+            peer_changes_queue: VecDeque::new(),
             peer_updates_sub: negotiation_subscriber,
         };
 
@@ -847,10 +852,9 @@ impl Peer<Stable> {
 
     /// Runs [`Task`]s which are scheduled for this [`Peer`].
     fn commit_scheduled_changes(&mut self) {
-        if !self.context.track_changes_queue.is_empty() {
-            while let Some(task) = self.context.track_changes_queue.pop_front()
-            {
-                self.context.pending_track_updates.push(task.clone());
+        if !self.context.peer_changes_queue.is_empty() {
+            while let Some(task) = self.context.peer_changes_queue.pop_front() {
+                self.context.pending_peer_updates.push(task.clone());
                 task.dispatch_with(self);
             }
 
@@ -869,7 +873,7 @@ impl Peer<Stable> {
     /// Should be called when negotiation was finished.
     fn negotiation_finished(&mut self) {
         self.context.is_known_to_remote = true;
-        self.context.pending_track_updates.clear();
+        self.context.pending_peer_updates.clear();
         self.commit_scheduled_changes();
     }
 }
@@ -889,13 +893,13 @@ impl<'a> PeerChangesScheduler<'a> {
     /// Provided [`TrackPatch`]s will be sent to the client on (re)negotiation.
     pub fn patch_tracks(&mut self, patches: Vec<TrackPatch>) {
         for patch in patches {
-            self.schedule_change(TrackChange::TrackPatch(patch));
+            self.schedule_change(PeerChange::TrackPatch(patch));
         }
     }
 
-    /// Schedules [`TrackChange::IceRestart`].
+    /// Schedules [`PeerChange::IceRestart`].
     pub fn restart_ice(&mut self) {
-        self.schedule_change(TrackChange::IceRestart);
+        self.schedule_change(PeerChange::IceRestart);
     }
 
     /// Schedules `send` tracks adding to `self` and `recv` tracks for this
@@ -938,30 +942,30 @@ impl<'a> PeerChangesScheduler<'a> {
         }
     }
 
-    /// Adds provided [`TrackChange`] to scheduled changes queue.
+    /// Adds provided [`PeerChange`] to scheduled changes queue.
     #[inline]
-    fn schedule_change(&mut self, job: TrackChange) {
-        self.context.track_changes_queue.push_back(job);
+    fn schedule_change(&mut self, job: PeerChange) {
+        self.context.peer_changes_queue.push_back(job);
     }
 
     /// Schedules [`Track`] addition to [`Peer`] receive tracks list.
     ///
     /// This [`Track`] will be considered new (not known to remote) and may be
     /// obtained by calling `Peer.new_tracks` after this scheduled
-    /// [`TrackChange`] will be applied.
+    /// [`PeerChange`] will be applied.
     #[inline]
     fn add_receiver(&mut self, track: Rc<MediaTrack>) {
-        self.schedule_change(TrackChange::AddRecvTrack(track));
+        self.schedule_change(PeerChange::AddRecvTrack(track));
     }
 
     /// Schedules [`Track`] addition to [`Peer`] send tracks list.
     ///
     /// This [`Track`] will be considered new (not known to remote) and may be
     /// obtained by calling `Peer.new_tracks` after this scheduled
-    /// [`TrackChange`] will be applied.
+    /// [`PeerChange`] will be applied.
     #[inline]
     fn add_sender(&mut self, track: Rc<MediaTrack>) {
-        self.schedule_change(TrackChange::AddSendTrack(track));
+        self.schedule_change(PeerChange::AddSendTrack(track));
     }
 }
 
@@ -1106,8 +1110,8 @@ pub mod tests {
         let peer = peer.set_remote_answer(String::new());
         assert_eq!(peer.context.receivers.len(), 1);
         assert_eq!(peer.context.senders.len(), 1);
-        assert_eq!(peer.context.pending_track_updates.len(), 2);
-        assert_eq!(peer.context.track_changes_queue.len(), 0);
+        assert_eq!(peer.context.pending_peer_updates.len(), 2);
+        assert_eq!(peer.context.peer_changes_queue.len(), 0);
         assert_eq!(rx.recv().unwrap(), PeerId(0));
     }
 
@@ -1156,7 +1160,7 @@ pub mod tests {
 
         assert_eq!(peer_id, PeerId(0));
         assert_eq!(changes.len(), 2);
-        assert!(peer.context.track_changes_queue.is_empty());
+        assert!(peer.context.peer_changes_queue.is_empty());
 
         let peer = peer.set_local_offer(String::new());
         peer.set_remote_answer(String::new());
@@ -1224,10 +1228,10 @@ pub mod tests {
 
         let mut track_patches_after: Vec<_> = peer
             .context
-            .pending_track_updates
+            .pending_peer_updates
             .iter()
             .filter_map(|t| {
-                if let TrackChange::TrackPatch(patch) = t {
+                if let PeerChange::TrackPatch(patch) = t {
                     Some(patch.clone())
                 } else {
                     None
@@ -1245,17 +1249,17 @@ pub mod tests {
     }
 
     #[test]
-    fn dedup_ice_restarts() {
+    fn ice_restart_dedupping_works() {
         let changes = vec![
-            TrackChange::IceRestart,
-            TrackChange::IceRestart,
-            TrackChange::IceRestart,
-            TrackChange::TrackPatch(TrackPatch {
+            PeerChange::IceRestart,
+            PeerChange::IceRestart,
+            PeerChange::IceRestart,
+            PeerChange::TrackPatch(TrackPatch {
                 id: TrackId(0),
                 is_muted: None,
             }),
-            TrackChange::IceRestart,
-            TrackChange::TrackPatch(TrackPatch {
+            PeerChange::IceRestart,
+            PeerChange::TrackPatch(TrackPatch {
                 id: TrackId(0),
                 is_muted: None,
             }),
@@ -1277,12 +1281,12 @@ pub mod tests {
             Rc::new(negotiation_sub),
         );
 
-        peer.context.pending_track_updates = changes;
+        peer.context.pending_peer_updates = changes;
 
         peer.dedup_ice_restarts();
 
-        let deduped_track_updates = peer.context.pending_track_updates;
+        let deduped_track_updates = peer.context.pending_peer_updates;
         assert_eq!(deduped_track_updates.len(), 3);
-        assert!(matches!(deduped_track_updates[1], TrackChange::IceRestart));
+        assert!(matches!(deduped_track_updates[1], PeerChange::IceRestart));
     }
 }